Add new colors, update logo, update folder name
