--- apps/chameleon/app/(dashboard)/dashboard/loading.tsx
@@ -1,21 +1,0 @@
-import { DashboardHeader } from '@/components/header'
-import { PostCreateButton } from '@/components/post-create-button'
-import { PostItem } from '@/components/post-item'
-import { DashboardShell } from '@/components/shell'
-
-export default function DashboardLoading() {
-  return (
-    <DashboardShell>
-      <DashboardHeader heading="Posts" text="Create and manage posts.">
-        <PostCreateButton />
-      </DashboardHeader>
-      <div className="divide-border-200 divide-y rounded-md border">
-        <PostItem.Skeleton />
-        <PostItem.Skeleton />
-        <PostItem.Skeleton />
-        <PostItem.Skeleton />
-        <PostItem.Skeleton />
-      </div>
-    </DashboardShell>
-  )
-}
