--- apps/chameleon/app/(dashboard)/dashboard/settings/loading.tsx
@@ -1,17 +1,0 @@
-import { CardSkeleton } from '@/components/card-skeleton'
-import { DashboardHeader } from '@/components/header'
-import { DashboardShell } from '@/components/shell'
-
-export default function DashboardSettingsLoading() {
-  return (
-    <DashboardShell>
-      <DashboardHeader
-        heading="Settings"
-        text="Manage account and website settings."
-      />
-      <div className="grid gap-10">
-        <CardSkeleton />
-      </div>
-    </DashboardShell>
-  )
-}
