+++ apps/chameleon/app/(loggedIn)/dashboard/billing/loading.tsx
@@ -1,0 +1,17 @@
+import { CardSkeleton } from '@/components/card-skeleton'
+import { DashboardHeader } from '@/components/header'
+import { DashboardShell } from '@/components/shell'
+
+export default function DashboardBillingLoading() {
+  return (
+    <DashboardShell>
+      <DashboardHeader
+        heading="Billing"
+        text="Manage billing and your subscription plan."
+      />
+      <div className="grid gap-10">
+        <CardSkeleton />
+      </div>
+    </DashboardShell>
+  )
+}
