+++ apps/chameleon/app/(loggedIn)/dashboard/todos/loading.tsx
@@ -1,0 +1,17 @@
+import { CardSkeleton } from '@/components/card-skeleton'
+import { DashboardHeader } from '@/components/header'
+import { DashboardShell } from '@/components/shell'
+
+export default function DashboardSettingsLoading() {
+  return (
+    <DashboardShell>
+      <DashboardHeader
+        heading="Settings"
+        text="Manage account and website settings."
+      />
+      <div className="grid gap-10">
+        <CardSkeleton />
+      </div>
+    </DashboardShell>
+  )
+}
