+++ apps/chameleon/tailwind.config.js
@@ -54,6 +54,45 @@ module.exports = {
           DEFAULT: 'hsl(var(--card))',
           foreground: 'hsl(var(--card-foreground))',
         },
+        green: {
+          50: '#effef3',
+          100: '#dafee5',
+          200: '#b8facb',
+          300: '#81f4a4',
+          400: '#2ce263',
+          500: '#1acd51',
+          600: '#0faa3e',
+          700: '#108535',
+          800: '#12692e',
+          900: '#115628',
+          950: '#033013',
+        },
+        pink: {
+          50: '#fef1f9',
+          100: '#fee5f4',
+          200: '#ffcaeb',
+          300: '#ff9fd9',
+          400: '#ff63bd',
+          500: '#ff299b',
+          600: '#f0127e',
+          700: '#d10562',
+          800: '#ad0750',
+          900: '#8f0c46',
+          950: '#580025',
+        },
+        blue: {
+          50: '#eafeff',
+          100: '#cbf9ff',
+          200: '#9ef2ff',
+          300: '#5be6ff',
+          400: '#29d4ff',
+          500: '#00b2e5',
+          600: '#008dc0',
+          700: '#03709b',
+          800: '#0d5a7d',
+          900: '#104b69',
+          950: '#033049',
+        },
       },
       borderRadius: {
         lg: 'var(--radius)',
